ensure 8-bit return values are treated correctly
the `mos_uopen` API was spotted to be not correctly handling the underlying `open_UART1` function’s return value, replacing it with the contents of the A register

new MOS 3 APIs `mos_extractnumber` and `mos_isdirectory` have been changed, as the underlying functions were returning a `BOOL` but there is no documenation/guidance on how ZDS treats booleans in the C calling convention.  for safety they are changed to use an 8-bit value instead of a BOOL, and the API implementations changed accordingly

+++ src/mos.c
@@ -3011,9 +3011,9 @@ UINT8 fat_lseek(FIL * fp, DWORD * offset) {
 	return f_lseek(fp, *offset);
 }
 
-int fat_getfree(const TCHAR * path, DWORD * clusters, DWORD * clusterSize) {
+UINT8 fat_getfree(const TCHAR * path, DWORD * clusters, DWORD * clusterSize) {
 	FATFS * fs = NULL;
-	int result;
+	UINT8 result;
 	if (clusters == NULL || clusterSize == NULL) {
 		return FR_INVALID_PARAMETER;
 	}
